Use literal suffix notation for timeout

+++ test/modules/MetricsValidator/src/Program.cs
@@ -62,7 +62,7 @@ namespace MetricsValidator
                 using (ModuleClient moduleClient = await ModuleClient.CreateFromEnvironmentAsync(settings))
                 using (MetricsScraper scraper = new MetricsScraper(new List<string> { "http://edgeHub:9600/metrics", "http://edgeAgent:9600/metrics" }))
                 {
-                    moduleClient.OperationTimeoutInMilliseconds = (uint)60000;
+                    moduleClient.OperationTimeoutInMilliseconds = 60000u;
 
                     Logger.LogInformation("Open Async");
                     await moduleClient.OpenAsync();
